Remove old hosted platform
