+++ src/Makefile
@@ -20,7 +20,7 @@ LDFLAGS+= -lopencm3_stm32f4 -lc -lnosys -flto
 LDFLAGS+= -T platforms/stm32f4-discovery.ld -nostartfiles
 all: tfi
 
-.elif ${PLATFORM} == "hosted2"
+.elif ${PLATFORM} == "hosted"
 CC=gcc
 CFLAGS+= -I. -std=c99 -Wall -O2 -ggdb -lrt -DSUPPORTS_POSIX_TIMERS
 CFLAGS+= -D TICKRATE=1000000 -D_POSIX_C_SOURCE=199309L -D_GNU_SOURCE
